feat(models): add category model and update resource item
- Create new Category model with basic fields
- Update ResourceItem model to use ForeignKey to Category
- Add DRF router setup for API endpoints in urls.py
