+++ category/models.py
@@ -1,0 +1,13 @@
+from django.db import models
+
+# Create your models here.
+
+
+class Category(models.Model):
+    name = models.CharField(max_length=200)
+    description = models.TextField(max_length=500)
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    def __str__(self):
+        return self.name
