+++ resource_item/models.py
@@ -1,17 +1,8 @@
 from django.db import models
 from django.contrib.auth.models import User
+from category.models import Category
 
 # Create your models here.
-
-
-class Category(models.Model):
-    name = models.CharField(max_length=200)
-    description = models.TextField(max_length=500)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    def __str__(self):
-        return self.name
 
 
 class ResourceItem(models.Model):
