+++ resource_item/urls.py
@@ -1,6 +1,16 @@
-from django.urls import path
-from . import views
+from django.urls import path, include
+from rest_framework.routers import DefaultRouter
+from .views import ResourceItemViewSet
+
+
+"""
+This file defines the available API endpoints for the ResourceItem model.
+Provides standard CRUD operations through DRF's DefaultRouter.
+"""
+
+router = DefaultRouter()
+router.register(r'', ResourceItemViewSet)
 
 urlpatterns = [
-    path('', views.resource_item, name='resource_item'),
+    path('', include(router.urls)),
 ]
